+++ huego_test.go
@@ -19,7 +19,6 @@ var hostname string
 var badHostname = "bad-hue-config"
 
 func init() {
-
 	hostname = os.Getenv("HUE_HOSTNAME")
 	username = os.Getenv("HUE_USERNAME")
 
@@ -436,7 +435,6 @@ func init() {
 		httpmock.RegisterResponder("PUT", fmt.Sprintf("http://%s/api%s", badHostname, p), httpmock.NewBytesResponder(200, response))
 		httpmock.RegisterResponder("DELETE", fmt.Sprintf("http://%s/api%s", badHostname, p), httpmock.NewBytesResponder(200, response))
 	}
-
 }
 
 func TestDiscoverAndLogin(t *testing.T) {
@@ -470,4 +468,58 @@ func Test_unmarshalError(t *testing.T) {
 	}
 	err := unmarshal([]byte(`not json`), s)
 	assert.NotNil(t, err)
+}
+
+// strPtr returns pointer to a string.
+func strPtr(val string) *string {
+	return &val
+}
+
+// boolPtr returns pointer to an boolean.
+func boolPtr(val bool) *bool {
+	return &val
+}
+
+// intPtr returns pointer to an integer.
+func intPtr(val int) *int {
+	return &val
+}
+
+// uint8Ptr returns pointer to an uint8.
+func uint8Ptr(val uint8) *uint8 {
+	return &val
+}
+
+// uint16Ptr returns pointer to an uint16.
+func uint16Ptr(val uint16) *uint16 {
+	return &val
+}
+
+// float32Ptr returns pointer to a float32.
+func float32Ptr(val float32) *float32 {
+	return &val
+}
+
+// float64Ptr returns pointer to a float64.
+func float64Ptr(val float64) *float64 {
+	return &val
+}
+
+// TestPointers tests all the above pointer methods
+func TestPointers(t *testing.T) {
+	s := "test"
+	b := true
+	i := int(123)
+	ui8 := uint8(123)
+	ui16 := uint16(123)
+	f32 := float32(1.234)
+	f64 := float64(1.234)
+
+	assert.True(t, *strPtr(s) == s)
+	assert.True(t, *boolPtr(b) == b)
+	assert.True(t, *intPtr(i) == i)
+	assert.True(t, *uint8Ptr(ui8) == ui8)
+	assert.True(t, *uint16Ptr(ui16) == ui16)
+	assert.True(t, *float32Ptr(f32) == f32)
+	assert.True(t, *float64Ptr(f64) == f64)
 }
