--- utils.go
@@ -1,36 +1,0 @@
-package huego
-
-// strPtr returns pointer to a string.
-func strPtr(val string) *string {
-	return &val
-}
-
-// boolPtr returns pointer to an boolean.
-func boolPtr(val bool) *bool {
-	return &val
-}
-
-// intPtr returns pointer to an integer.
-func intPtr(val int) *int {
-	return &val
-}
-
-// uint8Ptr returns pointer to an uint8.
-func uint8Ptr(val uint8) *uint8 {
-	return &val
-}
-
-// uint16Ptr returns pointer to an uint16.
-func uint16Ptr(val uint16) *uint16 {
-	return &val
-}
-
-// float32Ptr returns pointer to a float32.
-func float32Ptr(val float32) *float32 {
-	return &val
-}
-
-// float64Ptr returns pointer to a float64.
-func float64Ptr(val float64) *float64 {
-	return &val
-}
